extract connection logic from pg_model
