+++ src/connection/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "pg_model"
+name = "connection"
 version = "0.1.0"
 authors = ["Alex Dukhno <alex.dukhno@icloud.com>"]
 edition = "2018"
@@ -7,12 +7,20 @@ publish = false
 
 [dependencies]
 protocol = { path = "../protocol" }
+pg_model = { path = "../pg_model" }
 
+async-dup = "1.2.1"
+async-mutex = "1.4.0"
+#TODO temporal to have ability to run ./local/code_coverage.sh
+async-native-tls = { git = "https://github.com/alex-dukhno/async-native-tls.git", branch = "new-native-tls" }
+async-trait = "0.1.40"
+blocking = "1.0.0"
 bigdecimal = { version = "0.2.0", features = ["string-only"] }
 byteorder = "1.3.4"
-# TODO get rid off dependency
 sqlparser = { git = "https://github.com/ballista-compute/sqlparser-rs.git", branch = "main", features = ["bigdecimal"] }
+futures-lite = "1.10.0"
+log = "0.4.11"
 rand = "0.7"
 
 [dev-dependencies]
-rstest = "0.6.4"
+tempfile = "3.1.0"
