+++ src/node/Cargo.toml
@@ -14,6 +14,7 @@ path = "src/bin.rs"
 
 [dependencies]
 binder = { path = "../binder" }
+connection = { path = "../connection" }
 description = { path = "../description" }
 data_manager = { path = "../data_manager" }
 kernel = { path = "../kernel" }
