+++ src/query_executor/Cargo.toml
@@ -8,6 +8,7 @@ publish = false
 [dependencies]
 ast = { path = "../ast" }
 binary = { path = "../binary" }
+connection = { path = "../connection" }
 constraints = { path = "../constraints" }
 data_manager = { path = "../data_manager" }
 expr_eval = { path = "../expr_eval" }
